+++ src/FolkerKinzel.MimeTypes/Intls/MimeTypeSerializer.cs
@@ -1,0 +1,9 @@
+﻿using FolkerKinzel.MimeTypes.Intls.Parameters.Serializers;
+
+namespace FolkerKinzel.MimeTypes.Intls;
+
+internal static class MimeTypeSerializer
+{
+    
+}
+