+++ src/FolkerKinzel.MimeTypes/MimeTypeInfo_Properties.cs
@@ -132,7 +132,7 @@ public readonly partial struct MimeTypeInfo
     /// <returns>An appropriate file type extension for the <see cref="MimeTypeInfo"/> instance.</returns>
     /// <remarks>
     /// <para>
-    /// If no other file type extension could be found, <see cref="MimeCache.DefaultFileTypeExtension"/>
+    /// If no other file type extension could be found, <see cref="MimeCache.DEFAULT_EXTENSION_WITHOUT_PERIOD"/>
     /// is returned. <paramref name="includePeriod"/> specifies whether the period is included.
     /// </para>
     /// <para>