+++ src/FolkerKinzel.MimeTypes/MimeType_Methods.cs
@@ -271,7 +271,7 @@ public sealed partial class MimeType
     /// <returns>An appropriate file type extension.</returns>
     /// <remarks>
     /// <para>
-    /// If no other file type extension could be found, <see cref="MimeCache.DefaultFileTypeExtension"/>
+    /// If no other file type extension could be found, <see cref="MimeCache.DEFAULT_EXTENSION_WITHOUT_PERIOD"/>
     /// is returned. <paramref name="includePeriod"/> specifies whether the period is included.
     /// </para>
     /// <para>